services/ingestion_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
asyncpg==0.30.0
aiokafka==0.12.0
pydantic==2.11.5
pytz==2025.2
openai-whisper==20240930